yolo

+++ astro.config.mjs
@@ -1,7 +1,7 @@
 import { defineConfig } from 'astro/config';
 import tailwind from '@astrojs/tailwind';
 import react from '@astrojs/react';
-import { copyFileSync } from 'fs';
+import { copyFileSync, existsSync } from 'fs';
 
 // https://astro.build/config
 export default defineConfig({
@@ -12,7 +12,9 @@ export default defineConfig({
       name: 'copy-worker',
       hooks: {
         'astro:build:done': () => {
-          copyFileSync('worker.js', 'dist/worker.js');
+          if (existsSync('worker.js')) {
+            copyFileSync('worker.js', 'dist/worker.js');
+          }
         }
       }
     }
